+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
-  - Minor: Add maxLength attribute support to text and abstract custom fields
 </details>
+
+## v0.45.0
+  - Minor: Add maxLength attribute support to text and abstract custom fields
 
 ## v0.44.1
   - Patch: Import Select component in documentation